requirements.txt
# Core Flask dependencies
Flask==3.1.0
Flask-Cors==4.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# PDF and Image Processing
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.2.1
PyPDF2==3.0.1
pyheif==0.8.0

# Translation and Language Detection
googletrans==4.0.0-rc1
langdetect==1.0.9

# HTTP Requests
requests==2.28.2

# Environment variables
python-dotenv==1.0.1

# Utilities
typing_extensions==4.13.2
importlib_metadata==8.7.0
zipp==3.21.0
Mako==1.3.10
MarkupSafe==3.0.2

# Production server
gunicorn==21.2.0